Fix some more MSVC warnings, and make test_bandit_correctness faster.
The latter shouldn't waste a lot of time on outputting a progress
count to stderr now.

+++ src/main/test_monotonicity_bandit.cc
@@ -75,7 +75,7 @@ void test_with_bandits(std::vector<std::shared_ptr<election_method> > &
 			confident = true;
 		}
 
-		std::vector<std::pair<double, int> > so_far;
+		std::vector<std::pair<double, size_t> > so_far;
 
 		// Warning: No guarantee that this will contain the best until
 		// the method is finished; and no guarantee that it will contain
@@ -96,7 +96,7 @@ void test_with_bandits(std::vector<std::shared_ptr<election_method> > &
 			if (sims[k]->higher_is_better()) {
 				score = -score;
 			}
-			so_far.push_back(std::pair<double, int>(score, k));
+			so_far.push_back(std::pair<double, size_t>(score, k));
 		}
 		sort(so_far.begin(), so_far.end());
